# libs/runner-api-client/src/models/is-recoverable-response.ts
/* tslint:disable */

/**
 * Daytona Runner API
 * Daytona Runner API
 *
 * The version of the OpenAPI document: v0.0.0-dev
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface IsRecoverableResponse
 */
export interface IsRecoverableResponse {
  /**
   *
   * @type {boolean}
   * @memberof IsRecoverableResponse
   */
  recoverable?: boolean
}